[native pos] Refactor package structure of presto-spark-base
Move classes into more relevant package structure.

Packages created:
- http: For all http related classes (request, response, client)
- nativeprocess - (Native) process related classes
- property - Configuration classes
- shuffle - Shuffle related
- task - Native as well as non-native task related classes
- accesscontrol - Native as well as non-native Access Control related classes
