+++ presto-spark-base/src/main/java/com/facebook/presto/spark/PrestoSparkModule.java
@@ -104,18 +104,21 @@ import com.facebook.presto.server.SessionPropertyDefaults;
 import com.facebook.presto.server.TaskUpdateRequest;
 import com.facebook.presto.server.remotetask.RemoteTaskStats;
 import com.facebook.presto.server.security.ServerSecurityModule;
+import com.facebook.presto.spark.accesscontrol.PrestoSparkAccessControlChecker;
+import com.facebook.presto.spark.accesscontrol.PrestoSparkAuthenticatorProvider;
+import com.facebook.presto.spark.accesscontrol.PrestoSparkCredentialsProvider;
 import com.facebook.presto.spark.classloader_interface.SparkProcessType;
-import com.facebook.presto.spark.execution.BatchTaskUpdateRequest;
 import com.facebook.presto.spark.execution.PrestoSparkBroadcastTableCacheManager;
 import com.facebook.presto.spark.execution.PrestoSparkExecutionExceptionFactory;
-import com.facebook.presto.spark.execution.PrestoSparkLocalShuffleReadInfo;
-import com.facebook.presto.spark.execution.PrestoSparkLocalShuffleWriteInfo;
-import com.facebook.presto.spark.execution.PrestoSparkNativeTaskExecutorFactory;
-import com.facebook.presto.spark.execution.PrestoSparkTaskExecutorFactory;
+import com.facebook.presto.spark.execution.http.BatchTaskUpdateRequest;
 import com.facebook.presto.spark.execution.property.NativeExecutionConnectorConfig;
 import com.facebook.presto.spark.execution.property.NativeExecutionNodeConfig;
 import com.facebook.presto.spark.execution.property.NativeExecutionSystemConfig;
 import com.facebook.presto.spark.execution.property.NativeExecutionVeloxConfig;
+import com.facebook.presto.spark.execution.shuffle.PrestoSparkLocalShuffleReadInfo;
+import com.facebook.presto.spark.execution.shuffle.PrestoSparkLocalShuffleWriteInfo;
+import com.facebook.presto.spark.execution.task.PrestoSparkNativeTaskExecutorFactory;
+import com.facebook.presto.spark.execution.task.PrestoSparkTaskExecutorFactory;
 import com.facebook.presto.spark.node.PrestoSparkInternalNodeManager;
 import com.facebook.presto.spark.node.PrestoSparkNodePartitioningManager;
 import com.facebook.presto.spark.node.PrestoSparkQueryManager;
