+++ presto-spark-base/src/main/java/com/facebook/presto/spark/accesscontrol/PrestoSparkAccessControlCheckerExecution.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.spark.execution;
+package com.facebook.presto.spark.accesscontrol;
 
 import com.facebook.presto.Session;
 import com.facebook.presto.common.type.BooleanType;
