+++ presto-spark-base/src/main/java/com/facebook/presto/spark/accesscontrol/PrestoSparkAuthenticatorProvider.java
@@ -11,8 +11,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.spark.execution;
+package com.facebook.presto.spark.accesscontrol;
 
-public interface PrestoSparkShuffleWriteInfo
+import com.facebook.presto.spi.security.TokenAuthenticator;
+
+import java.util.Map;
+
+public interface PrestoSparkAuthenticatorProvider
 {
+    Map<String, TokenAuthenticator> getTokenAuthenticators();
 }
