+++ presto-spark-base/src/main/java/com/facebook/presto/spark/execution/http/BatchTaskUpdateRequest.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.spark.execution;
+package com.facebook.presto.spark.execution.http;
 
 import com.facebook.presto.server.TaskUpdateRequest;
 import com.fasterxml.jackson.annotation.JsonCreator;
