+++ presto-spark-base/src/main/java/com/facebook/presto/spark/execution/shuffle/PrestoSparkLocalShuffleInfoTranslator.java
@@ -17,10 +17,6 @@ import com.facebook.airlift.json.JsonCodec;
 import com.facebook.presto.Session;
 import com.facebook.presto.spark.classloader_interface.PrestoSparkShuffleReadDescriptor;
 import com.facebook.presto.spark.classloader_interface.PrestoSparkShuffleWriteDescriptor;
-import com.facebook.presto.spark.execution.PrestoSparkLocalShuffleReadInfo;
-import com.facebook.presto.spark.execution.PrestoSparkLocalShuffleWriteInfo;
-import com.facebook.presto.spark.execution.PrestoSparkShuffleReadInfo;
-import com.facebook.presto.spark.execution.PrestoSparkShuffleWriteInfo;
 import com.google.inject.Inject;
 
 import java.io.IOException;
