+++ presto-spark-base/src/main/java/com/facebook/presto/spark/execution/shuffle/PrestoSparkLocalShuffleReadInfo.java
@@ -11,10 +11,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.spark.execution;
+package com.facebook.presto.spark.execution.shuffle;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
+
+import java.util.List;
 
 import static com.google.common.base.MoreObjects.toStringHelper;
 import static java.util.Objects.requireNonNull;
@@ -26,33 +28,24 @@ import static java.util.Objects.requireNonNull;
  * this class, one should make sure to make corresponding changes in the above
  * C++ struct and its corresponding serde functionalities.
  */
-public class PrestoSparkLocalShuffleWriteInfo
-        implements PrestoSparkShuffleWriteInfo
+public class PrestoSparkLocalShuffleReadInfo
+        implements PrestoSparkShuffleReadInfo
 {
-    private final int numPartitions;
-    // Unique identifier for each query generated by Presto.
+    // Presto queryId which is unique to each query.
     private final String queryId;
-    // Unique identifier for each shuffle stage generated by Spark
-    private final int shuffleId;
+    // Partition ids which are supposed to be read by given shuffle read.
+    private final List<String> partitionIds;
     private final String rootPath;
 
     @JsonCreator
-    public PrestoSparkLocalShuffleWriteInfo(
-            @JsonProperty("numPartitions") int numPartitions,
+    public PrestoSparkLocalShuffleReadInfo(
             @JsonProperty("queryId") String queryId,
-            @JsonProperty("shuffleId") int shuffleId,
+            @JsonProperty("partitionIds") List<String> partitionIds,
             @JsonProperty("rootPath") String rootPath)
     {
-        this.numPartitions = numPartitions;
         this.queryId = requireNonNull(queryId, "queryId is null");
-        this.shuffleId = shuffleId;
+        this.partitionIds = requireNonNull(partitionIds, "partitionIds is null");
         this.rootPath = requireNonNull(rootPath, "rootPath is null");
-    }
-
-    @JsonProperty
-    public int getNumPartitions()
-    {
-        return numPartitions;
     }
 
     @JsonProperty
@@ -68,18 +61,17 @@ public class PrestoSparkLocalShuffleWriteInfo
     }
 
     @JsonProperty
-    public int getShuffleId()
+    public List<String> getPartitionIds()
     {
-        return shuffleId;
+        return partitionIds;
     }
 
     @Override
     public String toString()
     {
         return toStringHelper(this)
-                .add("numPartitions", numPartitions)
                 .add("queryId", queryId)
-                .add("shuffleId", shuffleId)
+                .add("partitionIds", String.join(", ", partitionIds))
                 .add("rootPath", rootPath)
                 .toString();
     }
