+++ presto-spark-base/src/main/java/com/facebook/presto/spark/execution/shuffle/PrestoSparkLocalShuffleWriteInfo.java
@@ -11,12 +11,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.spark.execution;
+package com.facebook.presto.spark.execution.shuffle;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-
-import java.util.List;
 
 import static com.google.common.base.MoreObjects.toStringHelper;
 import static java.util.Objects.requireNonNull;
@@ -28,24 +26,33 @@ import static java.util.Objects.requireNonNull;
  * this class, one should make sure to make corresponding changes in the above
  * C++ struct and its corresponding serde functionalities.
  */
-public class PrestoSparkLocalShuffleReadInfo
-        implements PrestoSparkShuffleReadInfo
+public class PrestoSparkLocalShuffleWriteInfo
+        implements PrestoSparkShuffleWriteInfo
 {
-    // Presto queryId which is unique to each query.
+    private final int numPartitions;
+    // Unique identifier for each query generated by Presto.
     private final String queryId;
-    // Partition ids which are supposed to be read by given shuffle read.
-    private final List<String> partitionIds;
+    // Unique identifier for each shuffle stage generated by Spark
+    private final int shuffleId;
     private final String rootPath;
 
     @JsonCreator
-    public PrestoSparkLocalShuffleReadInfo(
+    public PrestoSparkLocalShuffleWriteInfo(
+            @JsonProperty("numPartitions") int numPartitions,
             @JsonProperty("queryId") String queryId,
-            @JsonProperty("partitionIds") List<String> partitionIds,
+            @JsonProperty("shuffleId") int shuffleId,
             @JsonProperty("rootPath") String rootPath)
     {
+        this.numPartitions = numPartitions;
         this.queryId = requireNonNull(queryId, "queryId is null");
-        this.partitionIds = requireNonNull(partitionIds, "partitionIds is null");
+        this.shuffleId = shuffleId;
         this.rootPath = requireNonNull(rootPath, "rootPath is null");
+    }
+
+    @JsonProperty
+    public int getNumPartitions()
+    {
+        return numPartitions;
     }
 
     @JsonProperty
@@ -61,17 +68,18 @@ public class PrestoSparkLocalShuffleReadInfo
     }
 
     @JsonProperty
-    public List<String> getPartitionIds()
+    public int getShuffleId()
     {
-        return partitionIds;
+        return shuffleId;
     }
 
     @Override
     public String toString()
     {
         return toStringHelper(this)
+                .add("numPartitions", numPartitions)
                 .add("queryId", queryId)
-                .add("partitionIds", String.join(", ", partitionIds))
+                .add("shuffleId", shuffleId)
                 .add("rootPath", rootPath)
                 .toString();
     }
