+++ presto-spark-base/src/main/java/com/facebook/presto/spark/execution/task/ForNativeExecutionTask.java
@@ -11,11 +11,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.spark;
+package com.facebook.presto.spark.execution.task;
 
-import java.util.Map;
+import javax.inject.Qualifier;
 
-public interface PrestoSparkCredentialsProvider
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
+
+import static java.lang.annotation.ElementType.FIELD;
+import static java.lang.annotation.ElementType.METHOD;
+import static java.lang.annotation.ElementType.PARAMETER;
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
+
+@Retention(RUNTIME)
+@Target({FIELD, PARAMETER, METHOD})
+@Qualifier
+public @interface ForNativeExecutionTask
 {
-    Map<String, String> getCredentials();
 }
