+++ presto-spark-base/src/main/java/com/facebook/presto/spark/execution/task/PrestoSparkTaskExecution.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.spark.execution;
+package com.facebook.presto.spark.execution.task;
 
 import com.facebook.airlift.concurrent.SetThreadName;
 import com.facebook.presto.event.SplitMonitor;
